+++ diagnostics.h
@@ -1,7 +1,7 @@
 #ifndef DIAGNOSTICS_H
 #define DIAGNOSTICS_H
 
-void write_particles(float *, bool,char *);
-void write_particle(float *, bool);
+void write_particles(double *, bool,char *);
+void write_particle(double *, bool);
 
 #endif
